requirements.txt
django>=1.5.1
wheel==0.24.0

django-tastypie